fix: update Title model to include default value for tags
The Title model in dtos.py was updated to include a default value for the 'tags' field. This ensures that even if no tags are provided, an empty list will be assigned to the field. This change allows for better handling of the 'tags' field in the Title model.

Additionally, the TitleTagTable in model.py was modified to include the 'ondelete="CASCADE"' option for the foreign key constraints. This ensures that when a title or tag is deleted, the corresponding entry in the TitleTagTable will also be deleted.

These changes improve the functionality and data integrity of the Title module.